# docs/openapi.yaml
openapi: 3.0.3
info:
  title: eMAC — Drug Allergy Card API
  version: 1.0.0
  description: >
    ระบบบัตรแพ้ยาอิเล็กทรอนิกส์ (eMAC) — Node.js + Express + PostgreSQL.


    **Authentication:** login แบบ mock ที่ `POST /auth/session` (ระหว่างยังไม่ได้ client id จริงของ
    MOPH Provider ID) จะได้ session JWT → ส่งเป็น `Authorization: Bearer <token>` กับ endpoint ที่ต้องยืนยันตัวตน.


    **Tenant isolation:** ทุก endpoint ที่ต้อง auth จะกรองข้อมูลตาม `hospcode` ที่อยู่ใน token เท่านั้น
    (ห้ามส่ง hospcode ผ่าน query — จะถูกเพิกเฉย).


    **หมายเหตุ base path:** สเปกนี้ตั้ง server เป็น root ของโดเมน (`HTTP_BASE_PATH=` ว่าง).
    ถ้า deploy ใต้ subpath ให้แก้ `servers[].url` ให้ตรง (เช่น `.../emac`).
  contact:
    name: eMAC API
    url: https://emac.moph.go.th
servers:
  - url: https://emac.moph.go.th
    description: Production (root mount)
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /auth/providers:
    get:
      tags: [Auth]
      summary: (mock) รายชื่อโปรไฟล์ให้เลือก login
      description: เปิดเฉพาะโหมด mock (dev/POC). โหมด real จะคืน 404.
      security: []
      responses:
        '200':
          description: รายชื่อ provider จำลอง
          content:
            application/json:
              schema:
                type: object
                properties:
                  providers:
                    type: array
                    items: { $ref: '#/components/schemas/MockProfile' }
        '404':
          $ref: '#/components/responses/NotFound'
  /auth/session:
    post:
      tags: [Auth]
      summary: (mock) login ด้วย providerId → session JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [providerId]
              properties:
                providerId:
                  type: string
                  example: mock-pharm-001
      responses:
        '201':
          description: login สำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResult' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: ไม่ใช่บุคลากรการแพทย์ / โรงพยาบาลไม่อยู่ใน allowlist
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/me:
    get:
      tags: [Auth]
      summary: identity ปัจจุบันจาก session
      responses:
        '200':
          description: ตัวตนของผู้ใช้
          content:
            application/json:
              schema:
                type: object
                properties:
                  identity: { $ref: '#/components/schemas/Identity' }
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---------------- Patients ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: session JWT จาก `POST /auth/session`
...
  responses:
    BadRequest:
      description: request ไม่ถูกต้อง
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: ไม่มี/ token ไม่ถูกต้องหรือหมดอายุ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: ไม่พบทรัพยากร
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum: [BAD_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, UNPROCESSABLE, INTERNAL, SERVICE_UNAVAILABLE]
            message: { type: string }
            details: {}
            requestId: { type: string }
...
    Role:
      type: string
      enum: [doctor, pharmacist]
...
    MockProfile:
      type: object
      properties:
        providerId: { type: string, example: mock-pharm-001 }
        name: { type: string, example: ภญ. เจนจิรา วัฒนกุล }
        role: { type: string, example: pharmacist }
        hospcode: { type: string, example: '10670' }
        hospitalName: { type: string, example: โรงพยาบาลศิริราช }
        isMedicalPersonnel: { type: boolean }
    Identity:
      type: object
      properties:
        providerId: { type: string }
        hospcode: { type: string }
        hospitalName: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        name: { type: string }
        keyId: { type: string }
    LoginResult:
      type: object
      properties:
        token: { type: string, description: session JWT }
        expiresAt: { type: string, format: date-time }
        profile:
          type: object
          properties:
            providerId: { type: string }
            name: { type: string }
            role: { type: string }
            hospcode: { type: string }
            hospitalName: { type: string }
            keyId: { type: string }
...